Serve measured results, and add CI
Three changes that together move the project from "has measured numbers" to
"serves measured numbers, and cannot silently stop doing so".

Promotion. app/seed/promote.py reads published run records and writes the next
seed version with those methods marked MEASURED, carrying run id, dataset
checksum, raw artifact path and sample count. Seed files are the backend's
source of truth, so measured figures must reach them by a reproducible path
rather than by hand-editing JSON. support-ticket-routing is now 2 of 4
measured; its other two methods stay DEMO because no model provider is
connected, and the evidence rule therefore still declines to recommend.

The result is worth stating plainly: the measured methods do not clear the
problem's own requirement. Its default bar is 92% accuracy; measured, rules
reach 75.69% and traditional ML 89.50%. The demo figures claimed 91.7% and
94.6% — that both passed. They did not. The API now reports an empty passing
set, which is the correct answer and the kind of correction the project exists
to produce.

Types. latency_p50_ms was an int, which cannot represent a 0.108ms measurement
and displayed it as "0ms". Widened to float through the schema, model and
domain, and the frontend formatters now render sub-millisecond latency and
sub-cent cost rather than rounding them to zero.

CI. There was none. GitHub Actions now runs backend lint, format, strict mypy,
130 tests and a benchmark smoke run, plus a frontend gate that renders five
pages at two widths and verifies the evidence rule in a real browser against
the module the site ships. That gate is what stops the rule being weakened by
a frontend change that never touches the backend.

Two bugs found while wiring this up, both by the gate rather than by reading:
a module-level latency() helper collided with a local DOM binding of the same
name inside setupBenchmarkPage, and a regex rewriting the frontend data matched
only the mantissa of a scientific-notation cost, producing 1e-06e-06 and
breaking the module at parse time.

130 tests pass; ruff and strict mypy clean; the frontend gate passes.

+++ backend/app/domain/recommendation.py
@@ -60,7 +60,7 @@ class Requirements:
     """
 
     min_accuracy: float
-    max_latency_ms: int
+    max_latency_ms: float
     auditability_required: bool
     monthly_volume: int
 
@@ -77,7 +77,7 @@ class MethodCandidate:
     method_id: str
     complexity_rank: int
     accuracy: float
-    latency_ms: int
+    latency_ms: float
     cost_per_1k: float
     deterministic: bool
     auditable: bool
@@ -190,7 +190,7 @@ def evaluate_method(method: MethodCandidate, requirements: Requirements) -> Meth
     )
 
 
-def _selection_key(method: MethodCandidate) -> tuple[int, float, int, str]:
+def _selection_key(method: MethodCandidate) -> tuple[int, float, float, str]:
     """Ordering used to pick among passing methods.
 
     Complexity first (the product thesis), then the standard's tie-breaks: lower
